+++ include/mbgl/style/expression/expression.hpp
@@ -174,7 +174,8 @@ enum class Kind : int32_t {
     In,
     Within,
     Distance,
-    IndexOf
+    IndexOf,
+    Slice
 };
 
 class Expression {